[Handover] - chore: handover kpi refactor (#696)
